Move ProvVis up a level
