+++ context/app/static/js/components/ProvVis/tests/fixtures/complex-cwl.json
