+++ context/app/static/js/components/ProvVis/tests/fixtures/real-cwl.json
