+++ context/app/static/js/components/ProvVis/tests/fixtures/simple-cwl.json
